+++ lib/ZarrCore/src/Codecs/Codecs.jl
@@ -46,4 +46,8 @@ getCodec(::Type{<:Codec}, d::Dict) = error("Unimplemented")
 
 include("V3/V3.jl")
 
+@static if VERSION ≥ v"1.11"
+    include("public_names_codecs.jl")
+end
+
 end